+++ express-zod-api/src/result-helpers.ts
@@ -1,7 +1,7 @@
 import { Request } from "express";
 import createHttpError, { HttpError, isHttpError } from "http-errors";
 import * as R from "ramda";
-import { z } from "zod";
+import { globalRegistry, z } from "zod";
 import type { $ZodObject } from "zod/v4/core";
 import { NormalizedResponse, ResponseVariant } from "./api-response";
 import {
@@ -12,7 +12,6 @@ import {
 } from "./common-helpers";
 import { InputValidationError, ResultHandlerError } from "./errors";
 import { ActualLogger } from "./logger-helpers";
-import { getExamples } from "./metadata";
 import type { LazyResult, Result } from "./result-handler";
 
 export type ResultSchema<R extends Result> =
@@ -92,14 +91,12 @@ export const getPublicErrorMessage = (error: HttpError): string =>
 /** @see pullRequestExamples */
 export const pullResponseExamples = <T extends $ZodObject>(subject: T) =>
   Object.entries(subject._zod.def.shape).reduce<FlatObject[]>(
-    (acc, [key, schema]) =>
-      combinations(
-        acc,
-        getExamples(schema).map(R.objOf(key)),
-        ([left, right]) => ({
-          ...left,
-          ...right,
-        }),
-      ),
+    (acc, [key, schema]) => {
+      const { examples = [] } = globalRegistry.get(schema) || {};
+      return combinations(acc, examples.map(R.objOf(key)), ([left, right]) => ({
+        ...left,
+        ...right,
+      }));
+    },
     [],
   );