+++ express-zod-api/src/zod-plugin.ts
@@ -8,8 +8,8 @@
  * @desc Stores the argument supplied to .brand() on all schema (runtime distinguishable branded types)
  * */
 import * as R from "ramda";
-import { z } from "zod";
-import { getExamples, metaSymbol } from "./metadata";
+import { globalRegistry, z } from "zod";
+import { metaSymbol } from "./metadata";
 import { Intact, Remap } from "./mapping-helpers";
 import type {
   $ZodType,
@@ -108,7 +108,7 @@ const $EZBrandCheck = z.core.$constructor<$EZBrandCheck>(
 );
 
 const exampleSetter = function (this: z.ZodType, value: z.output<typeof this>) {
-  const examples = getExamples(this).slice();
+  const examples = globalRegistry.get(this)?.examples?.slice() || [];
   examples.push(value);
   return this.meta({ examples });
 };